Paths, logs and comments updated
