+++ package.json
@@ -23,7 +23,8 @@
     "react-query": "^3.39.2",
     "react-router-dom": "^6.3.0",
     "recoil": "^0.7.5",
-    "recoil-persist": "^4.2.0"
+    "recoil-persist": "^4.2.0",
+    "vite-tsconfig-paths": "^3.5.0"
   },
   "devDependencies": {
     "@types/react": "^18.0.17",
