+++ src/recoil/Atom.ts
@@ -4,8 +4,7 @@ import { recoilPersist } from "recoil-persist";
 const { persistAtom } = recoilPersist();
 
 export const authorizationState = atom({
-  key: "authorization", // unique ID (with respect to other atoms/selectors)
-  default: [], // default value (aka initial value)
+  key: "authorization",
+  default: [],
   effects_UNSTABLE: [persistAtom],
 });
-
